feat: Use zgenom for zsh packages
This change adds in zgenom as a zsh package/config manager.  It's still
using all the same omz plugins and so on, just in a lot more pleasant
way.

It also includes functionality to install zegnom on first run, if it's
not present, because I'll forget how to do that otherwise.